Made methods in MarkerDetectPar module-level functions
